[ET-VK][EZ] Refactor graph directory structure
I always get lost in the current structure, so we
1. introduce `containers/`, `ops/`, and `ops/impl/` to flatten the directory, and
2. fix some include header ordering.

```
[jorgep31415@devvm15882.vll0 /data/users/jorgep31415/fbsource/fbcode/executorch/backends/vulkan/runtime/graph (4f468a807)]$ tree
.
├── ComputeGraph.cpp
├── ComputeGraph.h
├── containers
│   ├── Constant.cpp
│   ├── Constant.h
│   ├── SharedObject.cpp
│   ├── SharedObject.h
│   ├── Types.cpp
│   ├── Types.h
│   └── Value.h
├── GraphConfig.h
└── ops
    ├── ExecuteNode.h
    ├── Functions.cpp
    ├── Functions.h
    ├── impl
    │   ├── Arithmetic.cpp
    │   ├── Arithmetic.h
    │   ├── Copy.cpp
    │   ├── Copy.h
    │   ├── Staging.cpp
    │   └── Staging.h
    ├── OperatorRegistry.cpp
    ├── OperatorRegistry.h
    └── PrepackNode.h

3 directories, 22 files
```

Differential Revision: [D54272391](https://our.internmc.facebook.com/intern/diff/D54272391/)

[ghstack-poisoned]
