+++ backends/vulkan/runtime/graph/ops/OperatorRegistry.cpp
@@ -6,8 +6,9 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-#include <executorch/backends/vulkan/runtime/graph/Functions.h>
-#include <executorch/backends/vulkan/runtime/graph/OperatorRegistry.h>
+#include <executorch/backends/vulkan/runtime/graph/ops/OperatorRegistry.h>
+
+#include <executorch/backends/vulkan/runtime/graph/ops/Functions.h>
 
 namespace at {
 namespace native {
