+++ backends/vulkan/runtime/graph/ops/PrepackNode.h
@@ -8,16 +8,13 @@
 
 #pragma once
 
-// @lint-ignore-every CLANGTIDY facebook-hte-BadMemberName
-
 #ifdef USE_VULKAN_API
 
 #include <ATen/native/vulkan/api/Context.h>
 #include <ATen/native/vulkan/api/Tensor.h>
 #include <ATen/native/vulkan/api/Types.h>
 
-#include <executorch/backends/vulkan/runtime/graph/Config.h>
-#include <executorch/backends/vulkan/runtime/graph/Value.h>
+#include <executorch/backends/vulkan/runtime/graph/containers/Value.h>
 
 namespace at {
 namespace native {
@@ -25,19 +22,26 @@ namespace vulkan {
 
 class ComputeGraph;
 
-struct SharedObject {
+/*
+ * Represents a single prepacking op in a ML model. In graph mode, ops will be
+ * implemented in a derived class that implements encode, which will implement
+ * encoding of shaders transferring necessary data (such as weights and biases)
+ * to the GPU.
+ */
+class PrepackNode {
   friend class ComputeGraph;
 
-  explicit SharedObject() = default;
+ public:
+  PrepackNode(ValueRef tref, ValueRef packed) : tref_{tref}, packed_{packed} {}
 
-  VkMemoryRequirements aggregate_memory_requirements;
-  VmaAllocationCreateInfo aggregate_create_info;
-  std::vector<ValueRef> users;
-  api::MemoryAllocation allocation;
+  virtual ~PrepackNode() = default;
 
-  void add_user(ComputeGraph* const graph, const ValueRef idx);
-  void allocate(ComputeGraph* const graph);
-  void bind_users(ComputeGraph* const graph);
+ protected:
+  ValueRef tref_;
+  ValueRef packed_;
+
+ public:
+  virtual void encode(ComputeGraph* graph) const = 0;
 };
 
 } // namespace vulkan
