+++ backends/vulkan/runtime/graph/ops/impl/Arithmetic.cpp
@@ -6,10 +6,11 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+#include <executorch/backends/vulkan/runtime/graph/ops/impl/Arithmetic.h>
+
 #include <ATen/native/vulkan/impl/Common.h>
 
-#include <executorch/backends/vulkan/runtime/graph/ops/Arithmetic.h>
-#include <executorch/backends/vulkan/runtime/graph/ops/Staging.h>
+#include <executorch/backends/vulkan/runtime/graph/ops/impl/Staging.h>
 
 namespace at {
 namespace native {
