+++ backends/vulkan/runtime/graph/ops/impl/Copy.cpp
@@ -6,7 +6,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-#include <executorch/backends/vulkan/runtime/graph/ops/Copy.h>
+#include <executorch/backends/vulkan/runtime/graph/ops/impl/Copy.h>
 
 namespace at {
 namespace native {
