+++ backends/vulkan/runtime/graph/ops/impl/Staging.h
@@ -10,9 +10,9 @@
 
 #ifdef USE_VULKAN_API
 
-#include <string.h>
-
 #include <executorch/backends/vulkan/runtime/graph/ComputeGraph.h>
+
+#include <string.h>
 
 namespace at {
 namespace native {
